requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
PyJWT==2.4.0 #
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.2.2  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.17.6  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.2.13  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-filter==21.1 #
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.1.0 #
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.22.1  # https://github.com/tfranzel/drf-spectacular
